Add files via upload
This is an ant colony simulator.
There are several types of ants in the colony: 
Harvesters gather food.
Warriors defend the colony against enemy attacks, which become more common as the colony's food supply increases.
Priests pray to the Ant Gods, increasing the colony's Divine Favor, which is then exchanged for favors from the Gods.
The Queen oversees the colony. If she dies, the colony dies with her.